requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
setuptools==78.1.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.1
waitress==3.0.2
Werkzeug==3.1.3
